Creation of new AlbumCollection component
