+++ src/App.jsx
@@ -1,8 +1,8 @@
+import React from 'react';
 import { data } from './data/data.js';
 import ArtistHeader from './components/ArtistHeader.jsx';
 import ActionButtons from './components/ActionButtons.jsx';
-import AlbumInfo from './components/AlbumInfo.jsx';
-import TrackList from './components/TrackList.jsx';
+import AlbumCollection from './components/AlbumCollection.jsx';
 
 function App() {
   const { profile, stats, discography } = data.artistUnion;
@@ -11,20 +11,18 @@ function App() {
 
   const getAlbumDetails = (data) => {
     const albums = data?.artistUnion?.discography?.albums.items || [];
-    const albumDetails =
-      albums.map((album) => {
-        const release = album?.releases?.items?.[0] || {};
-        return {
-          id: release.id || '',
-          songs: getSongsAlbum(album),
-        };
-      }) || [];
-    return albumDetails;
+    return albums.map((album) => {
+      const release = album?.releases?.items?.[0] || {};
+      return {
+        id: release.id || '',
+        songs: getSongsAlbum(album),
+      };
+    });
   };
 
   const getSongsAlbum = (album) => {
     const songsAlbum = album?.releases?.items?.[0]?.tracks?.items || [];
-    const songsDetails = songsAlbum.map((trackItem) => {
+    return songsAlbum.map((trackItem) => {
       const track = trackItem?.track || {};
       return {
         uid: trackItem?.uid || '',
@@ -34,8 +32,6 @@ function App() {
         duration: track.duration?.totalMilliseconds || '0',
       };
     });
-
-    return songsDetails;
   };
 
   const albumDetails = getAlbumDetails(data);
@@ -52,22 +48,12 @@ function App() {
       <div>
         <ActionButtons />
       </div>
-      {albums.map((albumGroup) =>
-        albumGroup.releases.items.map((album) => {
-          const songs = getSongsAlbum(album);
-          const albumDetail = albumDetails.find(detail => detail.id === album.id);
-          return (
-            <div key={album.id}>
-              <AlbumInfo album={album} />
-              {validateAlbumDetails && albumDetail ? (
-                <TrackList songs={albumDetail.songs} />
-              ) : (
-                <p>Album not found</p>
-              )}
-            </div>
-          );
-        })
-      )}
+      <AlbumCollection
+        albums={albums}
+        getSongsAlbum={getSongsAlbum}
+        albumDetails={albumDetails}
+        validateAlbumDetails={validateAlbumDetails}
+      />
     </div>
   );
 }
